+++ src/cfuncs/src/chisq.c
@@ -51,8 +51,8 @@ static PyObject *residuals(PyObject *self, PyObject *args){
     return NULL;
   }
   /* Get data and prior arrays size:                               */
-  dsize = PyArray_DIM(model,    0);
-  psize = PyArray_DIM(prioroff, 0);
+  dsize = (int)PyArray_DIM(model,    0);
+  psize = (int)PyArray_DIM(prioroff, 0);
   size[0] = dsize + psize;
 
   /* Initialize resuduals array:                                   */
@@ -123,7 +123,7 @@ static PyObject *chisq(PyObject *self, PyObject *args){
     return NULL;
   }
   /* Get data and prior arrays size:                               */
-  dsize = PyArray_DIM(model, 0);
+  dsize = (int)PyArray_DIM(model, 0);
 
   /* Calculate model chi-squared:                                  */
   for(i=0; i<dsize; i++){
